Reduce size of font for text longer than 120 chars
